% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{addTo_wrapper}
\alias{addTo_wrapper}
\title{Add to number from 1 to x, in C}
\usage{
addTo_wrapper(x)
}
\arguments{
\item{x}{an int}
}
\value{
an int
}
\description{
Add to number from 1 to x, in C
}
\examples{
addTo_wrapper(100)
addTo_wrapper(10)
}
